% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_simplex.R
\name{apply_simplex}
\alias{apply_simplex}
\title{Apply Simplex
Function to apply simplex to a data frame}
\usage{
apply_simplex(E = 2:5, lib, pred, dat_in)
}
\arguments{
\item{E}{Embedding dimensions; defaults to 2:5}

\item{lib}{index of library values}

\item{pred}{index of predicted values}

\item{dat_in;}{data used in the function}
}
\description{
Apply Simplex
Function to apply simplex to a data frame
}
