% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_simplex_list.R
\name{apply_simplex_list}
\alias{apply_simplex_list}
\title{Apply simplex list function
Function to apply simplex to list of sampled data for values of X, Y, and Z}
\usage{
apply_simplex_list(E, lib, pred, samp_ts)
}
\arguments{
\item{E}{Embedding dimensions}

\item{lib}{Index of library values}

\item{pred}{Index of prediction values}

\item{samp_ts}{Output from sample_data function}
}
\description{
Apply simplex list function
Function to apply simplex to list of sampled data for values of X, Y, and Z
}
