% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dat.R
\name{load_dat}
\alias{load_dat}
\title{Load dat
Function to either load or simulate data}
\usage{
load_dat(ctl_in, ncores)
}
\arguments{
\item{ctl_in}{Ctl list in; output from mtheory_ctl function}

\item{ncores}{Number of cores}
}
\description{
Load dat
Function to either load or simulate data
}
