% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make filename ctl
Function to generate filenames from ctl}
\usage{
make_filename(ctl_in)
}
\arguments{
\item{ctl_in}{Ctl from mtheory_ctl function}
}
\description{
Make filename ctl
Function to generate filenames from ctl
}
