% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multivariate}
\alias{multivariate}
\title{Multivariate function
Function to evaluate multivariate rEDM models}
\usage{
multivariate(ctl_in, dat_in, max_lag, base_column, add_column, target_column)
}
\arguments{
\item{ctl_in}{Ctl list; used to define lib and pred in block_lnlp}

\item{dat_in}{Data frame of sampled values; from run_simplex[[1]]}

\item{max_lag}{Maximum lag value for make_block}

\item{base_column}{Base columns to include in multivariable prediction}

\item{add_column}{Additional columns to evaluate}

\item{target_column}{Target column to predict}
}
\description{
Multivariate function
Function to evaluate multivariate rEDM models
}
