% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simplex.R
\name{run_simplex}
\alias{run_simplex}
\title{Run Simplex
Function to simulate data, sample data at some frequency, and
apply simplex methods to evaluate predictive ability}
\usage{
run_simplex(ctl_in, ncores = 6)
}
\arguments{
\item{ctl_in}{Generated by mtheory_ctl function}

\item{ncores}{Number of cores}
}
\description{
Run Simplex
Function to simulate data, sample data at some frequency, and
apply simplex methods to evaluate predictive ability
}
