% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample Data
Function to sample data}
\usage{
sample_data(samp_freq = 50, data_in = samps)
}
\arguments{
\item{samp_freq}{Frequency to sample data}

\item{data_in}{Data input to the function}
}
\description{
Sample Data
Function to sample data
}
