% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results.R
\name{show_results}
\alias{show_results}
\title{Show simplex results
Function to show simplex results. Sampled data, embedding dimensions, prediction
decays and nonlinearity}
\usage{
show_results(samp_input, nrun)
}
\arguments{
\item{samp_input}{Output from run_simplex}

\item{nrun}{Iteration to sample}
}
\description{
Show simplex results
Function to show simplex results. Sampled data, embedding dimensions, prediction
decays and nonlinearity
}
