% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_list.R
\name{simplex_list}
\alias{simplex_list}
\title{Apply simplex to a list
Function to apply simplex to}
\usage{
simplex_list(input_df, E = 2:10, lib, pred)
}
\arguments{
\item{input_df}{Input data frame}

\item{E}{Dimensions}

\item{lib}{Library indices, no default}

\item{pred}{Prediction indices, no default}
}
\description{
Apply simplex to a list
Function to apply simplex to
}
